GB: generate fixed sequences for constant shift amounts
Shifting by an runtime amount requires a loop, this can be replaced by
a series of shifts and rotates if we know the rotation amount during
codegen.

This is a substantial codegen improvement but also shouldn't be
necessary. Certain compares calculate the sign bit using a 7-bit shift,
we could avoid the mask if they correctly produces a 1-bit rotate.