GB: support hram routines, move rodata into ROM1
hram routines are necessary for implementing a DMA wait without touching
the bus.

Maybe I can be smarter with the ROM usage and move all the transient
initialization data into its own bank (which would be switched out by
the runtime). This would reduce the burden on the C programmer.