GB: add support for %hi and %lo relocation specifiers
The following assembly is now valid:
```
ld a, c
add %lo memory
ld a, b
add %hi memory
```

This saves a register compared to a plain 16-bit relocation. This is now
supported by the assembler but is not currently produced by the compiler.

This is an unfortunately big change since I can no-longer use the
default MCExpr implementation.