GB: add support for atomics
Atomics are necessary for any data-structures that need to be shared
with an interrupt handler (eg. a task-queue). Interestingly, since the
hardware is single-threaded, everything MUST be lock-free -- no
libatomic calls allowed here.

Unfortunately, since there is no way to query the current global
interrupt state (and no atomic swap), the __sync libcall implementations
must either: unconditionally reenable interrupts OR perform a non-atomic
swap of the interrupt-enable register.

For now, I have opted to do a non-atomic swap of the interrupt-enable
register... This means that interrupts are no-longer allowed to touch
this value -- otherwise they may invoking a race-condition.